# [Enhancement] safe guard to jni string